+++ src/__tests__/local-state/export.ts
@@ -3,7 +3,7 @@ import { print } from 'graphql/language/printer';
 
 import { Observable } from '../../util/Observable';
 import { ApolloLink } from '../../link/core';
-import ApolloClient from '../..';
+import { ApolloClient } from '../..';
 import { InMemoryCache } from '../../cache/inmemory';
 
 describe('@client @export tests', () => {