+++ src/__tests__/optimistic.ts
@@ -6,7 +6,7 @@ import gql from 'graphql-tag';
 import { mockSingleLink } from '../__mocks__/mockLinks';
 import { MutationQueryReducersMap } from '../core/types';
 import { Subscription } from '../util/Observable';
-import ApolloClient from '../';
+import { ApolloClient } from '../';
 import { addTypenameToDocument, stripSymbols } from '../utilities';
 import { InMemoryCache, makeReference } from '../cache/inmemory';