+++ src/react/context/ApolloContext.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ApolloClient from '../../ApolloClient';
+import { ApolloClient } from '../../ApolloClient';
 
 export interface ApolloContextValue {
   client?: ApolloClient<object>;